python-services/extraction/requirements.txt
# FastAPI Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Azure Document Intelligence
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
azure-identity==1.19.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Utilities
python-dotenv==1.0.1
structlog==24.4.0

# Health Check
prometheus-client==0.21.1
